Returned to older, working version

+++ cw04/zad3/main.c
@@ -13,13 +13,26 @@ static int TYPE;
 static char* prog_name;
 pid_t child_pid = 0;
 int child_recieved = 0;
+int child_end = 0;
 int parent_sent = 0;
 int parent_recieved = 0;
+int parent_end = 0;
 
 void show_help () {
     printf("Usage: %s <Num of signals> <Type>\n"
+            " TYPE 1 – parent sends signals without waiting\n"
+            "      2 – parent waits for answer for each signal\n"
+            "      3 – processes send real time signals\n"
            , prog_name);
     exit(EXIT_FAILURE);
+}
+
+void set_signal(int signum, void* handler) {
+    struct sigaction act_SIGUSR1;
+    act_SIGUSR1.sa_handler = handler;
+    sigfillset(&act_SIGUSR1.sa_mask);
+    act_SIGUSR1.sa_flags = 0;
+    sigaction(signum, &act_SIGUSR1, NULL);
 }
 
 void child_exit() {
@@ -36,24 +49,13 @@ void child_SIGUSR1(int signum) {
 }
 
 void child_SIGUSR2(int signum) {
-    kill(getppid(), SIGUSR2);
-    child_exit();
+    child_end = 1;
 }
 
 void child_setup() {
-    struct sigaction act_SIGUSR1;
-    act_SIGUSR1.sa_handler = child_SIGUSR1;
-    sigfillset(&act_SIGUSR1.sa_mask);
-    sigdelset(&act_SIGUSR1.sa_mask, SIGUSR1);
-    act_SIGUSR1.sa_flags = 0;
-    sigaction(SIGUSR1, &act_SIGUSR1, NULL);
-
-    struct sigaction act_SIGUSR2;
-    act_SIGUSR2.sa_handler = child_SIGUSR2;
-    sigfillset(&act_SIGUSR2.sa_mask);
-    sigdelset(&act_SIGUSR2.sa_mask, SIGUSR2);
-    act_SIGUSR2.sa_flags = 0;
-    sigaction(SIGUSR2, &act_SIGUSR2, NULL);
+    set_signal(SIGUSR1, child_SIGUSR1);
+    set_signal(SIGUSR2, child_SIGUSR2);
+    child_recieved = 0;
 
     printf("Child (%d) setup done\n", getpid());
 }
@@ -71,33 +73,33 @@ void parent_exit() {
 
 void parent_SIGUSR1(int signum) {
     parent_recieved += 1;
-    if (parent_recieved < L) {
-        parent_sent += 1;
-        kill(child_pid, SIGUSR1);
-    } else {
-        parent_exit();
+
+    if (TYPE == 3) {
+        if (parent_recieved < L) {
+            parent_sent += 1;
+            kill(child_pid, SIGUSR1);
+        } else {
+            parent_end = 1;
+        }
     }
 }
 
 void parent_setup() {
-    struct sigaction act_SIGUSR1;
-    act_SIGUSR1.sa_handler = parent_SIGUSR1;
-    sigemptyset(&act_SIGUSR1.sa_mask);
-    sigaddset(&act_SIGUSR1.sa_mask, SIGUSR2);
-    act_SIGUSR1.sa_flags = 0;
-    sigaction(SIGUSR1, &act_SIGUSR1, NULL);
+    set_signal(SIGUSR1, parent_SIGUSR1);
+    parent_sent = 0;
+    parent_recieved = 0;
 
     printf("Parent (%d) setup done\n", getpid());
-    kill(child_pid, SIGUSR1);
-    parent_sent = 1;
 }
 
 int main (int argc, char** argv) {
 
+    // Read args
     prog_name = argv[0];
     if (argc == 3) {
         L = atoi(argv[1]);
         TYPE = atoi(argv[2]);
+        if (TYPE < 1 || TYPE > 3) show_help();
     } else {
         show_help();
     }
@@ -108,12 +110,33 @@ int main (int argc, char** argv) {
         perror("Cannot make a fork");
         exit(EXIT_FAILURE);
     } else if (child_pid == 0) {
+        // Child fork
         child_setup();
     } else {
+        // Parent fork
         parent_setup();
+        switch (TYPE) {
+            case 1:
+                while (parent_sent < L) {
+                    kill(child_pid, SIGUSR1);
+                    parent_sent += 1;
+                }
+                parent_end = 1;
+                break;
+            case 2:
+                break;
+            case 3:
+                kill(child_pid, SIGUSR1);
+                parent_sent = 1;
+                break;
+        }
     }
 
-    // Do nothing
-    while (1) pause();
+    while (1) {
+        if (child_end) child_exit();
+        if (parent_end) parent_exit();
+        pause();
+    }
+
     return EXIT_SUCCESS;
 }
